This patch does several things:
Change opcode_size into get_instruction, which returns you a copy of the
instruction or None if the instruction isn't known. This allows callers to catch
the error of not knowing an instruction.  This extra copy was happening already
in the callers of opcode_size.  Sadly because of the borrow checker, I can't
figure out how to eliminate this copy.

Change opcode_dispatch to instruction_dispatch.

Change the opcode generated code's pyunit test to not assert the exact contents.