Change the debugger to not require the emulator to store debugging information inside of it.
When detecting changes to memory the debugger simulates the next instruction
with a read-only version of the emulator that intercepts writes to memory.  This
not only allows a decoupling of the debugging information from the emulator, but
allows the debugger to stop before executing an instruction that modifies an
address of interest.